+++ addOns/scripts/src/main/java/org/zaproxy/zap/extension/scripts/automation/actions/RemoveScriptAction.java
@@ -36,6 +36,8 @@ public class RemoveScriptAction extends ScriptAction {
     private static final List<String> DISABLED_FIELDS =
             Arrays.asList(
                     ScriptJobDialog.SCRIPT_ENGINE_PARAM,
+                    ScriptJobDialog.SCRIPT_INLINE_PARAM,
+                    ScriptJobDialog.SCRIPT_IS_INLINE_PARAM,
                     ScriptJobDialog.SCRIPT_FILE_PARAM,
                     ScriptJobDialog.SCRIPT_TARGET_PARAM);